+++ autogalaxy/analysis/aggregator/__init__.py
@@ -1,2 +1,2 @@
 from autogalaxy.imaging.model.aggregator import ImagingAgg
-from autogalaxy.analysis.aggregator.aggregator import InterferometerAgg
+from autogalaxy.interferometer.model.aggregator import InterferometerAgg
