+++ autogalaxy/interferometer/interferometer.py
