+++ autogalaxy/interferometer/model/aggregator.py
