+++ autogalaxy/plot/fit_interferometer_plotters.py
@@ -1,7 +1,7 @@
 import autoarray.plot as aplt
 from autoarray.plot import fit_interferometer_plotters
 
-from autogalaxy.fit.fit_interferometer import FitInterferometer
+from autogalaxy.interferometer.fit_interferometer import FitInterferometer
 from autogalaxy.plot.mat_wrap.lensing_mat_plot import MatPlot1D
 from autogalaxy.plot.mat_wrap.lensing_mat_plot import MatPlot2D
 from autogalaxy.plot.mat_wrap.lensing_visuals import Visuals1D
