+++ test_autogalaxy/imaging/model/test_aggregator_imaging.py
