+++ test_autogalaxy/interferometer/model/test_aggregator_interferometer.py
